backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.1
langchain==0.1.5
langchain-community>=0.0.17
langchain-google-genai==0.0.6
google-generativeai>=0.3.2
supabase==2.3.0
requests==2.31.0
gitpython==3.1.40
python-multipart==0.0.6
beautifulsoup4==4.12.3
markdown==3.5.2
psycopg2-binary==2.9.9
